+++ lib/index.js
@@ -1,76 +1,28 @@
-const axios = require("axios");
 const terminal = require("terminal-kit").terminal;
-const { table } = require("table");
 const chalk = require("chalk");
 const _ = require("lodash");
 
+const { buildRow, data } = require("./table");
+const { chancesOver, explainGame } = require("./messages");
+const { fetchWord } = require("./fetchWord");
+
 let word = "";
 let wordArr = [];
-const data = {
-  totalChances: 6,
-  elapsedChances: 0,
-  guessedWords: [],
-};
-
-function explainGame() {
-  terminal("Guess the WORDLE in six tries\n");
-  terminal(
-    "Each guess must be a valid five-letter word. \nHit the enter button to submit.\n"
-  );
-  terminal(
-    "After each guess, the color of the tiles will change to show how close your guess was to the word.\n"
-  );
-}
-
-async function startGame() {
-  word = await fetchWord();
-  // word = "issue";
-  word = word.toUpperCase();
-  // console.log(word);
-  wordArr = word.split("");
-  buildRow();
-  getInput();
-}
 
 // Calls the methods only when run through command line
 if (require.main === module) {
-  explainGame();
   startGame();
+  getInput();
 }
 
-async function buildRow() {
-  const defaultMatrix = [
-    [" ", " ", " ", " ", " "],
-    [" ", " ", " ", " ", " "],
-    [" ", " ", " ", " ", " "],
-    [" ", " ", " ", " ", " "],
-    [" ", " ", " ", " ", " "],
-    [" ", " ", " ", " ", " "],
-  ];
-  const emptyRow = [" ", " ", " ", " ", " "];
-  let matrix = [];
-  if (data.guessedWords.length === 0) {
-    matrix = defaultMatrix;
-  } else {
-    for await (const guessedWord of data.guessedWords) {
-      matrix.push(guessedWord);
-    }
-    if (
-      data.guessedWords.length > 0 &&
-      data.guessedWords.length < data.totalChances
-    ) {
-      let num = data.totalChances - data.guessedWords.length;
-      for (let i = 0; i < num; i++) {
-        matrix.push(emptyRow);
-      }
-    }
-  }
-  return drawTable(matrix);
-}
-
-function drawTable(tableData) {
-  console.log("\n");
-  console.log(table(tableData));
+async function startGame() {
+  explainGame();
+  word = await fetchWord();
+  // word = "issue";
+  word = word.toUpperCase();
+  console.log(word);
+  wordArr = word.split("");
+  buildRow();
 }
 
 async function getInput() {
@@ -83,17 +35,12 @@ async function getInput() {
   }).promise;
 }
 
-function chancesOver() {
-  terminal.red("\nSorry you had only " + data.totalChances + " chances");
-  terminal.red("\nYou lost! The word was " + word);
-}
-
 async function checkWord(wordGuess) {
   wordGuess = wordGuess.toUpperCase();
   const guessWord = await handleWordGuess(wordGuess);
   if (data.elapsedChances === data.totalChances) {
     buildRow();
-    chancesOver();
+    chancesOver(word);
     return process.exit();
   } else if (wordGuess.length !== 5) {
     terminal.bold.red("\nIt should be a 5 letter word\n");
@@ -189,37 +136,15 @@ async function handleWrongGuess() {
   if (data.elapsedChances !== data.totalChances) {
     await getInput();
   } else {
-    chancesOver();
+    chancesOver(word);
   }
-}
-
-async function fetchWord() {
-  let response;
-  const alphabets = "abcdefghijklmnopqrstuvwxyz";
-  const randomChar = alphabets[Math.floor(Math.random() * alphabets.length)];
-  try {
-    response = await axios.get(
-      "https://api.datamuse.com/words?sp=" + randomChar + "????&max=15"
-    );
-  } catch (error) {
-    console.log(error);
-    console.log("Something went wrong. Try again in few minutes");
-  }
-  let random = response.data[Math.floor(Math.random() * response.data.length)];
-  return random.word;
 }
 
 module.exports = {
   data,
-  explainGame,
-  startGame,
-  buildRow,
-  drawTable,
   getInput,
   checkWord,
   storeWord,
-  chancesOver,
   handleWordGuess,
   handleWrongGuess,
-  fetchWord,
 };
